Update promotions for Elementor One
