+++ src/php/class-plugin.php
@@ -160,9 +160,9 @@ class Plugin {
 		// Importers.
 		new Plugins_Import_Manager();
 		new Files_Import_Manager();
-		
+
 		// Initialize promotions.
-		new Promotions\Elementor_Pro();
+		new Promotions();
 	}
 
 	/**
